Geen crash meer als X niet goed geladen is. Jordi heeft een aantal mooie modellen gekregen. Draaipunt van camera is nu 0,n,0

+++ 3DEngine/Code/Camera.cpp
@@ -228,7 +228,7 @@ void Camera::useJoystickInput(DIJOYSTATE2 joystickState)
 {
 	if (joystickState.rgdwPOV[0] < 4294967295 )
 	{
-		float degrees = joystickState.rgdwPOV[0] / 100;
+		float degrees = (float)joystickState.rgdwPOV[0] / 100;
 		float radians = RADIANS(degrees);
 		xMovement = sin(radians) * 0.1f;
 		yMovement = cos(radians) * 0.1f;
